优化提示词和bug修复
1.修复impression个人印象错误的参数
2.优化提示词让bot对于@请求必须回复
3.其他关于impression传入错误参数的修复
4.尝试修复i686的构建

+++ src/logic/fate.rs
@@ -37,7 +37,7 @@ pub async fn fate(ctx: Context) -> Result<()> {
 
     if let Some(user_id) = sender.user_id {
         debug!(user_id = user_id, "开始获取用户印象并进行 AI 解签");
-        let impression = get_impression(user_id as i32).await;
+        let impression = get_impression(user_id).await;
         debug!(user_id = user_id, impression = ?impression, "成功获取用户印象");
 
         let prompt = vec![
